Remove functionality for calibration and preprocessing of raw data from sensors
No need for calibration for all sensors, as most calibrations are done once at production, and only offsets need to be adjusted. Removal of separated preprocessing step, as no need for holding both the raw and processed values (only need one, and we want the processed values to be kept).